move files from root to /app
