+++ shared/app/command-line.desktop.js
