+++ shared/app/global-errors/container.js
@@ -1,9 +1,9 @@
 // @flow
 import GlobalError from './index'
 import {connect} from 'react-redux'
-import {globalErrorDismiss} from '../constants/config'
+import {globalErrorDismiss} from '../../constants/config'
 
-import type {TypedState} from '../constants/reducer'
+import type {TypedState} from '../../constants/reducer'
 
 export default connect(
   (state: TypedState) => ({
