+++ shared/app/globals.native.js
