+++ shared/app/main-shared.native.js
@@ -1,21 +1,21 @@
 // @flow
-import DumbSheet from './dev/dumb-sheet'
-import Push from './push/push.native'
+import DumbSheet from '../dev/dumb-sheet'
+import Push from '../push/push.native'
 import React, {Component} from 'react'
-import RenderRoute from './route-tree/render-route'
-import loadPerf from './util/load-perf'
-import hello from './util/hello'
-import {Box} from './common-adapters/index'
-import {bootstrap} from './actions/config'
+import RenderRoute from '../route-tree/render-route'
+import loadPerf from '../util/load-perf'
+import hello from '../util/hello'
+import {Box} from '../common-adapters/index'
+import {bootstrap} from '../actions/config'
 import {connect} from 'react-redux'
 import {debounce} from 'lodash'
-import {getUserImageMap, loadUserImageMap} from './util/pictures'
-import {initAvatarLookup, initAvatarLoad} from './common-adapters/index.native'
-import {listenForNotifications} from './actions/notifications'
-import {persistRouteState, loadRouteState} from './actions/platform-specific.native'
-import {navigateUp, setRouteState} from './actions/route-tree'
+import {getUserImageMap, loadUserImageMap} from '../util/pictures'
+import {initAvatarLookup, initAvatarLoad} from '../common-adapters/index.native'
+import {listenForNotifications} from '../actions/notifications'
+import {persistRouteState, loadRouteState} from '../actions/platform-specific.native'
+import {navigateUp, setRouteState} from '../actions/route-tree'
 
-import type {TypedState} from './constants/reducer'
+import type {TypedState} from '../constants/reducer'
 
 type Props = {
   dumbFullscreen: boolean,
