+++ shared/app/main.ios.js
