+++ shared/app/nav.js.flow
@@ -2,9 +2,9 @@
 import {Component} from 'react'
 import {Map} from 'immutable'
 
-import type {RouteRenderStack} from './route-tree/render-route'
-import type {Reachability} from './constants/types/flow-types'
-import type {Tab} from './constants/tabs'
+import type {RouteRenderStack} from '../route-tree/render-route'
+import type {Reachability} from '../constants/types/flow-types'
+import type {Tab} from '../constants/tabs'
 
 export type Props = {
   appFocused: boolean,
