+++ shared/desktop/yarn-helper/index.js
@@ -3,6 +3,8 @@
 import buildCommands from './build'
 import electronComands from './electron'
 import fontCommands from './font'
+import fs from 'fs'
+import path from 'path'
 import visdiff from './visdiff'
 import {execSync} from 'child_process'
 import {padEnd} from 'lodash'
@@ -27,13 +29,39 @@ const commands = {
   },
   postinstall: {
     code: () => {
-      // Inject dummy module
-      exec('node make-shim net')
-      exec('node make-shim tls')
-      exec('node make-shim msgpack')
+      // Inject dummy modules
+      makeShims(['net', 'tls', 'msgpack'])
     },
     help: 'all: install global eslint. dummy modules',
   },
+}
+
+function makeShims(shims) {
+  shims.forEach(shim => {
+    const root = path.resolve(__dirname, '..', '..', 'node_modules', shim)
+
+    try {
+      fs.mkdirSync(root)
+    } catch (_) {}
+
+    try {
+      fs.writeFileSync(
+        path.join(root, 'package.json'),
+        `{
+  "main": "index.js"
+}
+`
+      )
+    } catch (_) {}
+
+    try {
+      fs.writeFileSync(
+        path.join(root, 'index.js'),
+        `module.exports = null // Generated shim-module
+`
+      )
+    } catch (_) {}
+  })
 }
 
 function exec(command, env, options) {
