+++ shared/index.android.js
@@ -4,6 +4,6 @@ import 'core-js/es6/object' // required for babel-plugin-transform-builtin-exten
 import 'core-js/es6/array' // required for emoji-mart in RN Android
 import 'core-js/es6/string' // required for emoji-mart in RN Android
 import 'core-js/es6/map' // required for FlatList in RN Android
-import {load} from './index.native'
+import {load} from './app/index.native'
 
 load()
