+++ shared/index.ios.js
@@ -1,5 +1,5 @@
 // @flow
 // React-native tooling assumes this file is here, so we just require our real entry point
-import {load} from './index.native'
+import {load} from './app/index.native'
 
 load()
